flow [nfc]: Convert $PropertyType to indexed-access syntax
Such a nicer syntax!  This is available to us now that we've
upgraded Prettier:
  https://github.com/zulip/zulip-mobile/pull/5393#issuecomment-1151612888
and Flow before that:
  https://github.com/zulip/zulip-mobile/pull/5398#discussion_r890444882

We'll similarly convert $ElementType in the next commit.

Done mostly with a crude search-and-replace:

    $ perl -i -0pe '
          s/\$PropertyType<(.*?),\s*(.*?)>/${1}[$2]/gs
        ' src/**/*.js types/**.js.flow flow-typed/**/*.js

That doesn't behave right in the presence of nesting; but there was
only one case of that, so just fixed it up by hand.