unread: Fix sort on marking 1:1 PMs unread
This was an `Array#sort` call without arguments, which falls into
a classic JS gotcha: it sorts by string value, even for numbers,
so 100 sorts before 2.

Fix the logic, and add a test.

There are similar-looking calls `….sort()` in the corresponding
code for stream messages in `unreadModel.js`.  But those do
correctly sort by numeric value, because the receivers of those
method calls are `Immutable.List` values, and those have the
good semantics.

Happily, the stream-messages code also already has a test
exercising the sort.  So if we had had a bug like this there,
we'd have caught it thanks to the test.  (In fact the way I found
this bug is that I was writing the comparable logic for group PMs,
and did write a test exercising the sort, and it failed.)