unread [nfc]: Specify types on creating a fresh Immutable.Map
Without this, the type of `messages` in this line:

    newlyUnreadState = newlyUnreadState.map(e => e.map(messages => messages.sort()));

is `empty`, meaning there's effectively no type-checking on it.

Conversely, having added these types, we can delete the rather
verbose-looking type annotations in the `mergeWith` call, and
everything there remains properly typed.

I'd learned this lesson before when working elsewhere in this code,
and even wrote it down in our style guide:
  https://github.com/zulip/zulip-mobile/blob/main/docs/style.md#immutable-provide-type

But then apparently I'd forgotten it later:
  https://github.com/zulip/zulip-mobile/pull/4790#discussion_r696980667