modelTypes [nfc]: For edit_history, FL <118, finish choosing null over void
We decided to go with `null` for the pre-118 case instead of
undefined/missing. Update a comment to say that, and update a type
cast to not say that servers might send `null`, since they won't.
(That type cast, and another like it, still isn't correct; we'll fix
another issue soon.)