ComposeMenu [nfc]: Use our own new type definition for attachments
Soon, we'd like to also use the `insertAttachments` prop not just
for react-native-document-picker payloads, but also for
react-native-image-picker payloads. Better to define our own type
that's meant to accommodate both kinds, rather than, e.g., using
react-native-document-picker's type and fudging image-picker
payloads to flow through that type.

While we're at it, trim down the type to just what
`insertAttachments` currently uses: the url [1] and a nullable name
for the file. The `type` property in particular smells like
something that might not be easy to make uniform across
document-picker and image-picker payloads; there's no comment
linking to a spec that says what the possible values of `type` mean,
for example.

[1] renamed from "uri"; see style guide