api.tryGetFileTemporaryUrl: Give null if href doesn't parse into URL object
One of this function's callers already doesn't make the call if
`href` fails to parse (with auth.realm) into a URL object: that's
the caller in messageLinkPress in src/message/messagesActions.js;
see the control flow there.

So any functional change would happen in the two callers in the
lightbox action-sheet code: one where we download the image, and one
where we go to share it with another app. I've marked those callers
in the diff by adding a jsdoc to the functions they appear in.

In a recent commit, we added a useEffect that navigates away from
the lightbox immediately on entering it, with navigation.goBack(),
if the URL string doesn't parse. That makes the effect of an
unparseable URL string on these two callers much less visible,
because of how much less likely they are to call with an invalid
`href`.

I suppose it's theoretically still possible: the user might open the
action sheet during the navigate-back animation, or perhaps the
navigate-back action fails because of some rare bug. So, to study
how the behavior changes within that small realm of possibility…

Before this commit, we'd expect an `href` that doesn't parse to be
rejected by the server when api.getFileTemporaryUrl makes its
request. Digging down to `apiFetch`, the URL for the request would
be equivalent to:

  new URL(`/api/v1/${href}`, auth.realm)

And if `href` is a string that doesn't parse to a URL with
auth.realm…then it must be something pretty weird that servers
probably won't recognize as trying to reach the documented endpoint
  /api/v1/user_uploads/{realm_id_str}/{filename}
.

At this commit, the function just doesn't let us make those requests
in the first place, and returns the same value, null.

So, no big user-visible effects are expected. It'll come with the
benefits of reducing network traffic, and it'll change the app's
behavior as witnessed by the server. So, not marking NFC.