StreamListScreen: Use Screen component, like other screens
I don't see a special reason for this component to avoid using the
Screen component. It didn't use Screen back when it was part of a
tab rendered by MainTabsScreen, and that made sense. But in
9d11e787f, it took on this new role as a regular stack screen on the
AppNavigator, and the norm is to use Screen for those.

I see just two functional changes, and they both seem like
improvements:

- The bottom inset is now padded, so now the list doesn't overlap
  with any notches, native UI, etc.
- Like other screens, this one now shows the OfflineNotice.