ComposeMenu: Remove a negative margin; don't make touch targets overlap
With padding 12 (and no negative margin), each button's touch target
is 12 + 24 + 12 = 48, and they don't overlap with each other. (The
24 is the icon's width.)

We try to keep touch targets 48px minimum so they're not too hard to
hit.

This makes them take up a bit more horizontal space, but that's OK
here -- it doesn't shrink the message text input down to zero width,
on my iPhone, and also we're about to move these buttons into a row
below the message text input.