api.updateMessageFlags [nfc]: Rename from messagesFlags
This matches the name in the OpenAPI description.  It also makes a
more grammatical phrase, and reads more naturally and informatively
at call sites.

(The old name dates from before we had OpenAPI descriptions; it
seems to be based on the URL path, which was a reasonable choice at
the time.)