api: Make ServerSettings a bit more "cooked"
We don't currently use these three properties of the response, but
if we do, we'd want them in these more convenient forms for passing
around. (We will want to use zulip_version soon, though, for #5102,
"Don't allow connecting to ancient servers.")

Not marking as NFC just because this will cause fresh errors if
zulip_version doesn't parse. There's no reason to suspect it
wouldn't parse, though.