AuthScreen [nfc]: Make realm-url expressions less of a mouthful
And put them in a consistent format, `serverSettings.realm_url`,
that'll be easy to handle when we eventually convert this component
to a function component.

Some of these intermediate declarations are placed such that there's
now a yield between when we read the current realm-URL prop and when
we use it:

- In `beginWebAuth`, there's an `await webAuth.generateOtp()`

- In `handleNativeAppleAuth`, there are these:
    await webAuth.generateRandomToken()
    await AppleAuthentication.signInAsync(…)
    await webAuth.generateOtp()

But that's fine and NFC, because the `serverSettings` route param
doesn't change and isn't supposed to change. Since that constraint
wasn't yet explicit, add a comment in the route-params type. (The
UX/security reason given in the comment has always applied, but the
constraint conveniently helps us verify this commit as NFC.)