accountsReducer: Invariant that ACCOUNT_SWITCH's index is in bounds
If this invariant isn't met, we'll end up with an AccountsState that
has `undefined` for its first item, and that would crash things up
pretty quickly.

Fortunately, half of the invariant already holds: `action.index`
won't be negative. Here's how we know that for both of
accountSwitch's two callsites:

- In the callsite in AccountPickScreen, the index comes from React
  Native's `FlatList`: it's meant to be the index for one of the
  account items to render. I don't think React Native would ever
  make that negative.

- In the callsite in narrowToNotification, it's the index passed to
  an array's `forEach` callback (the `state.accounts` array), so
  that can't be negative.

Of course, here in the reducer, we're using `action.index` as an
index into the `state.accounts` array as it stands at dispatch time.
If dispatchers are doing the same thing, there shouldn't be a
problem with the other half of the invariant: that `action.index`
isn't too high.

There seems to be just one place we fall short there: unfortunately,
the caller in AccountPickScreen was yielding between the time it
read `state.accounts` and dispatched the action, with an unexplained
setTimeout.

…So, remove that setTimeout, since in theory `state.accounts` could
change during those few milliseconds, such that `action.index` is
too high. Since in this case the reducer would add an `undefined` to
the end of `state.accounts`, add a migration. I tested the
AccountPickScreen manually after this change, and switching accounts
worked fine.