license: Strip copyright notice from LICENSE; update the one in README
This makes the whole LICENSE file an exact copy of the license,
which helps some software recognize that this is indeed the
Apache 2.0 license.  In particular, the `licensee` gem, which
is what GitHub uses to identify what license to display.

There's already a copyright notice in the README.
Update that, while we're here.