ios [nfc]: Rename AppDelegate.mm from ….m, following template
I opened the file in Xcode, used "Navigate > Reveal in Project Navigator",
and then long-clicked it there to get a UI to rename the file.
That moved the file itself, and made these small changes in the
Xcode project file.

This follows a change in the template app on the way to RN v0.68:
  facebook/react-native@4fb49cd55 Rename AppDelegate.m to AppDelegate.mm in the app template

which was done as preparation for the New Arch-related changes we'll
follow next.  That commit has a bunch of additional churn and new mess
in the Xcode project file, which apparently Xcode did automatically
when that change's author tried opening the project; but happily that
didn't happen for me.