ZLPConstants: Convert to Swift
Soon, we'd like to expose a URL to pass to Linking.openURL to bring
the user to the system-level notification settings for the Zulip
app.

I think I was caught by a red herring when getting such a URL on a
simulator running iOS 16. I was using the URL given by the global
variable UIApplicationOpenNotificationSettingsURLString, new in iOS
15.4 --
  https://developer.apple.com/documentation/uikit/uiapplicationopennotificationsettingsurlstring?language=objc
-- and openURL was bringing me to the home page of the Settings app
instead of the Zulip-specific notification settings.

At first, I thought it was failing because I wasn't using the
Swift-only UIApplication.openNotificationSettingsURLString --
  https://developer.apple.com/documentation/uikit/uiapplication/4013180-opennotificationsettingsurlstrin
-- new in iOS 16.0 to replace that
UIApplicationOpenNotificationSettingsURLString, which in the Swift
doc is marked as deprecated:
  https://developer.apple.com/documentation/uikit/uiapplicationopennotificationsettingsurlstring

But then I did this Swift conversion and tried out the iOS 16+
thing, and it gave the same URL with the same bad result. :shrug:
Now my theory is that something is busted with how Linking.openURL
works on the simulator. I tried with my iPhone and it opened to the
notification settings, as expected.

Anyway, might as well keep the work I did to convert this small
file, since Swift is the future. After reading docs and forum
discussions, including RN's official doc
  https://reactnative.dev/docs/0.68/native-modules-ios#exporting-swift
, here's what I did:

- Translate the Objective-C implementation, in ZLPConstants.m, to
  Swift, and rename the file from .m to .swift. Make sure it still
  has a PBXFileReference that appears in the ZulipMobile PBXGroup
  and in the "Compile Sources" build phase of the ZulipMobile
  target.
- Remove the .h file, its PBXFileReference, and the one place its
  PBXFileReference appeared: as a member of the ZulipMobile
  PBXGroup.
- Add a ZLPConstantsBridge.m, to register the Swift implementation
  with React Native. Make sure this new file has a PBXFileReference
  that appears in the ZulipMobile PBXGroup and in "Compile Sources".
- We already have a "bridging header",
  ios/ZulipMobile-Bridging-Header.h. Possibly we only need its first
  line, but I'm not sure, and it seems to work as-is.