account actions: Have removeAccount take an Identity, not an index
This encapsulates away from dispatchers the detail that
`state.accounts` is implemented as an array.

Now, callers have to be more expressive about what account they want
to remove: not "the one at index 2", but "the one with this realm
and this email".

That should fix the following bug, which is admittedly hard to
trigger but is still a bug. Here's what I did on Android at v27.196:

- Tap an account's trashcan icon

- Very quickly, before the confirmation dialog can show up and
  prevent you from doing so, tap elsewhere on the card for the
  account, so the app switches to that account

- Press "Confirm" on the "Remove account" dialog

- See that an account was removed, but it wasn't the one mentioned
  in the dialog