NotificationsScreen: Label per-account settings as such
I could imagine someone seeing

- Notifications when offline
- Notifications when online
- Stream notifications

and wrongly thinking that they were device-level instead of
account-level. So, group these per-account settings together with a
header that identifies the account they apply to, with a link to the
pick-account screen in case the user wants to set these values on a
different account.