msglist: Give error alert when tapped link has URL that didn't parse
Just before this commit, if an `href` fails to parse, it would do so
in the `new URL` call that's meant to pass a URL object to
`getNarrowFromLink` (the line with the comment "TODO: Use parsedUrl,
below").

The error would propagate up to messageLinkPress's caller, in
handleWebViewOutboundEvent, and from there to a React event handler,
with the result (I expect) that the app wouldn't crash, but it also
wouldn't give any useful feedback to the user.

So, give some useful feedback to the user. The issue is filed as
"Fail early (and tell user) on a message-link tap when the URL
doesn't parse" (#5518) . I guess we have been failing early, we just
haven't been telling the user.

Fixes: #5518