ZulipVersion: Add `toJSON` method
We'll use this to make a JSON report of notification problems
without having to implement special replacer logic for
`ZulipVersion`s.

It seems fairly likely that we've been including ZulipVersion
objects in payloads sent to Sentry. For those, I expect this change
will have them use the raw version string instead of the whole JSON
stringified object. That seems like an improvement.