ZLPConstants: Return `true` in `requiresMainQueueSetup`
When writing the previous commit, I noticed that UIApplication comes
from UIKit. React Native's doc makes it sound like we should return
`true` here when our module "require[s] access to UIKit"; see code
comment.

(When this module first appeared, in 4828ecf32, the comment
correctly said that we didn't use UIKit. The use of UIApplication
was added later, in 7471f6e1e.)

I'm not aware of a functional change this would cause. The doc
  (https://reactnative.dev/docs/0.68/native-modules-ios#exporting-constants)
says:

> in the future your module may be initialized on a background
> thread unless you explicitly opt out with
> `+ requiresMainQueueSetup:`

but (with "in the future") that doesn't sound like a behavior change
we'd see now. Anyway, the features supported by this module (viewing
a message list, and linking to system settings for Zulip
notifications) still work fine, as tested manually on my iPhone 13
Pro running iOS 16.1.