NotificationListener [nfc]: Have listenIOS take a params object
This will be useful soon when we have listenIOS use a new
NativeEventEmitter we've written, instead of relying on the emitter
@react-native-community/push-notification-ios gives us, which is
buggy.

The `addListener` method of our new NativeEventEmitter will be typed
such that the passed event name and handler are checked against each
other. Now that listenIOS has a single param with both the name and
the handler, it'll be easy to write it so it expects one or another
kind of name/handler pair, by typing its param as a union of object
types. Then the new caller can pass a value that flows through one
branch of the union, to the new NativeEventEmitter, without changing
any type constraints on the existing callers.