deps: Temporarily use fork for @react-navigation/stack
(Much like we did for @react-navigation/bottom-tabs in c0b17bd32.)

To get chrisbobbe/react-navigation@a588dcee9, which I've published
to NPM in @zulip/react-navigation-stack@5.14.10-0.zulip.1.

It cherry-picks react-navigation/react-navigation@5a1987708, which
fixes #5486, as noted there:
  https://github.com/zulip/zulip-mobile/issues/5486#issuecomment-1261138257

So this fix will become unnecessary when we're on React Navigation
6; that's issue #4936.

Tested the bugfix on my iPhone 13 Pro running iOS 16.1, with and
without "Debug with Chrome". The bug was fixed, and back navigation
worked as expected, both with the swipe gesture and the back button.

Fixes: #5486