messagesActions [nfc]: Handle unexpected narrow links more explicitly
This should give the same behavior, just without an odd run-around
in the case where isNarrowLink gives true but
getNarrowFromNarrowLink gives false.

Before, in that case, we'd call api.tryGetFileTemporaryUrl, which
would give null because of the unexpected pathname ("/" instead of
something that starts with "/user_uploads/"). We'd recognize that
`null` and take the intended action, calling
openLinkWithUserPreference.

Refactor and add some comments; and for that case, just call
openLinkWithUserPreference without going through an API method.