presence tests: Remove one that expects fallback we don't need
We removed the explicit fallback code this was testing in b94544258.
As it turns out, the way we use Immutable.Map (new in that commit's
same PR) effectively handles the ancient-server behavior as well as
the explicit fallback was, and so this test still passes. But that's
incidental and not something we need to guarantee.