presence [nfc]: Modelify statusFromPresence, as getPresenceOnlyStatusForUser
This new function takes the whole PresenceState as an argument,
instead of a small fragment of it, which will help us as we add
new wrinkles to that state for #5669.

It also returns null for missing data, rather than picking a value;
it's then up to the caller to choose how to handle that.