mute: Don't offer mute/unmute in action sheet when not subscribed
The concept of muting or unmuting a stream can only exist when
the user is subscribed in the first place; without a subscription,
there isn't even a place in the data model for the muted/unmuted
state to live.

For a topic, the data model doesn't rule it out but it wouldn't
mean anything: you don't get unreads or notifications when you're
not subscribed.