compose: Focus message input on quote-and-reply
Note that we have logic in `handleMessageFocus` to respond to focus
on the message input by auto-focusing the *topic* input in some
cases. That would be annoying if it happened here: you do a
quote-and-reply, and suddenly the topic input is focused?(!)

But that topic auto-focusing should be defeated in all cases where
it was active. In particular, since the previous commit, the topic
input won't be empty; the logic in `handleMessageFocus` will see
that and skip the auto-focus.

Fixes: #5718