ServerCompatBanner [nfc]: Remove "notes on known breakage" comment
Before we had a practice of refusing to connect to ancient servers
(see kMinAllowedServerVersion), compatibility would break at those
ancient versions in idiosyncratic ways, and it made sense to keep
track of those ways in a comment like this.

Now, instead of dropping ancient-server compatibility in one or
another area of the app, we just refuse to connect at all, breaking
the user's experience in a uniform and predictable way (along with a
prompt to the user to get their server upgraded). So we don't need
to keep this list around anymore. (Also, note that 1.8 is well below
kMinAllowedServerVersion, so that specific item hasn't been accurate
for a while.)