android: Update targetSdkVersion to 33 (Android 13)
This change will be required in order to upload new releases to the
Play Store, effective 2023-08-31 (or 2023-11-01 if we request an
extension).

This change causes Android 13 and later to apply to our app a number
of behavior changes introduced in that version, documented here:
  https://developer.android.com/about/versions/13/behavior-changes-13

The "Granular media permissions" section stood out as potentially
relevant -- until we remembered that, in PR #5618, we upgraded
react-native-image-picker such that on Android 13+ we use Android's
PhotoPicker API. With that API, the app learns about just the images
that the user selects, in a separate media picker, so our app code
doesn't need direct access to whole lists of media created by other
apps. That's what this note in that section of the doc is about:

> Note: If your app only needs to access images, photos, and videos,
> consider using the photo picker instead of declaring the
> READ_MEDIA_IMAGES and READ_MEDIA_VIDEO permissions.

Fixes: #5453