sentry: Disable Sentry for now on Android
This is a workaround for #5757, the crashes we've started to
regularly see as many users' devices have started taking the
Google Play system update dated 2023-08.

See the (long) chat thread here:
  https://chat.zulip.org/#narrow/stream/48-mobile/topic/Android.20crashes/near/1640893

and issue threads reported in Android's and Sentry's trackers:
  https://issuetracker.google.com/issues/300840851
  https://github.com/getsentry/sentry-unity/issues/1444

Our issue for re-enabling Sentry on Android is #5766.

This commit was cherry-picked from 29cea899d, part of v27.213,
with tweaks to mention #5766.