unread [nfc]: Fix a confusingly inaccurate comment
This meant to refer to *topics* that are in the new state
but not the old, not individual messages.  Fix that; and add
a parenthetical explaining a bit more, too.

Also cut the "too" on saying we sort here.  It made sense
before 09eee4eac; but with that change, there's no other
sort going on in the code this comment is discussing.

Thanks to Chris for pointing out this comment's confusingness:
  https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/flutter.3A.20unreads.20model/near/1647712