i18n: Offer Latvian (lv) and Tagalog (tl) in the UI
I've just gone through Transifex, and I found two languages that
it's time to offer in the UI! Transifex reports that, for Latvian
and Tagalog, mobile.json is 24.4% and 5.1% translated, respectively.
Those are both above 5%. So, following our process described in
src/settings/languages.js, it's time to add them to the UI.

Thanks as always to our kind volunteer translators.