CustomProfileFields: Use styled UserItem for (unknown user)
Since #5790 last month, UserItem has been prepared to gracefully
handle users we don't know about. Use that, instead of
special-casing with a plain ZulipTextIntl, which we'd been doing
since 2022-06 (0a108312c) when CustomProfileFields was added.

The change won't be visible unless a given user object is actually
absent; that'll start happening when we do #5808.

See screenshots:
  https://github.com/zulip/zulip-mobile/pull/5809#issuecomment-1868051890