msglist: Get message sender's name from user, not message, when available
But if it's not in user data, just fall back to sender_full_name on
the message, since that seems better than showing "(unknown user)".

Like in the previous commit, the main motivation for this is that
our new helper functions that will power the upcoming "Firstname
Lastname (guest)" feature, #5804, need a User object.

In the tests for the "content" WebView inbound event, tweak the
example message so that its sender is the self-user. That's so the
tested code can access the sender's User object, since the self-user
is already included in the test's background data.