android build: Upgrade Android Gradle Plugin to 7.2.2
Release notes here:
  https://developer.android.com/build/releases/past-releases/agp-7-2-0-release-notes

Done because the upcoming `yarn upgrade` would otherwise cause a
build failure with a message that looks similar to one that
reportedly was resolved by upgrading to AGP 7.2.0:
  https://stackoverflow.com/a/74808149
and it turns out that our error is also resolved by upgrading. In
our case we upgraded to the latest 7.2.x, which right now is 7.2.2.

Our error message had this:

  gradle Error: Can't determine type for tag
  '<macro name="m3_comp_assist_chip_container_shape">?attr/shapeAppearanceCornerSmall</macro>'