settings: For subtitle of selected language, repeat canonical self-name
It looks odd to have an item where the title and subtitle are
slightly different ways of naming the same thing. The canonical
self-name is the one we've chosen by following organizations
with a broad global contributor base like Wikipedia, so use that.

(We could instead just omit the subtitle, but if we did, then all
the rows in the list would have two lines of text except the
selected row which would have just one. And that non-uniformity
might look odd.)