ios: Disable Flipper
Our build times could probably have benefited from this for a while
now; we never got into the habit of actually using Flipper.

Flipper is a debugging tool. I believe its code has not been
included in release builds.

Anyway, disable it now, since things would otherwise be busted with
recent Xcode, as mentioned in a comment.

Done by commenting out the relevant line in the Podfile, and running
`yarn` (resulting in a `pod install`) and doing a debug build.